+++ src/index.ts
@@ -44,6 +44,7 @@ export {
   type InlineChipProps,
   type InlineChipTone,
 } from './InlineChip'
+export { Lightbox, type LightboxProps } from './Lightbox'
 export { Link, type LinkProps, type LinkVariant } from './Link'
 export { IconButton, type IconButtonProps, type IconButtonVariant } from './IconButton'
 export { Kbd, type KbdProps } from './Kbd'